Move ResourceIcons to shared folder
The same icons will be used both in the resource page and in search
results.
Refs #435.
