+++ app/screens/shared/resource-icons/ResourceIcons.js
