+++ app/screens/shared/resource-icons/ResourceIcons.spec.js
