+++ app/screens/shared/resource-icons/index.js
@@ -1,0 +1,3 @@
+import ResourceIcons from './ResourceIcons';
+
+export default ResourceIcons;
