+++ app/screens/shared/resource-icons/resource-icons.less
@@ -1,0 +1,13 @@
+.resource-icons {
+  color: @gray;
+  margin-bottom: 8px;
+
+  .glyphicon {
+    font-size: 16px;
+  }
+
+  .text {
+    margin-left: 4px;
+    margin-right: 10px;
+  }
+}
