+++ app/screens/shared/resource-list/resource-list.less
@@ -50,20 +50,6 @@
       padding: 20px;
     }
 
-    .resource-icons {
-      color: @gray;
-      margin-bottom: 8px;
-
-      .glyphicon {
-        font-size: 16px;
-      }
-
-      .text {
-        margin-left: 4px;
-        margin-right: 10px;
-      }
-    }
-
     h4 {
       margin-top: 0;
       margin-bottom: 0;
