+++ app/screens/shared/shared.less
@@ -2,4 +2,5 @@
 @import './comment-form/comment-form.less';
 @import './date-header/date-header.less';
 @import './favorite-button/favorite-button.less';
+@import './resource-icons/resource-icons.less';
 @import './resource-list/resource-list.less';
